pulumi/postgres_read_replicas/requirements.txt
psycopg2-binary==2.8.6
pulumi==2.17.0
pulumi-aiven==3.3.2
Faker==5.6.5